CreateNewDB  implemented

+++ IndustrialUnit.Model/Model/Valve.cs
@@ -8,7 +8,7 @@ namespace IndustrialUnit.Model.Model
     public int? Id { get; set; }
     public string ItemType { get; set; }
     public string Operation { get; set; }
-    public decimal? Size { get; set; }
+    public int? Size { get; set; }
     public string ConnectionType { get; set; }
     public string Supplier { get; set; }
     public string Manufacturer { get; set; }
